% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylmergeR-package.R
\docType{package}
\name{methylmergeR-package}
\alias{methylmergeR}
\alias{methylmergeR-package}
\title{methylmergeR: CpG Methylation Cross-Pipeline Consensus Filtering}
\description{
Tools for filtering, comparing, and merging CpG methylation calls across multiple alignment/calling pipelines (Bismark, BWA-meth, Biscuit, ENCODE), including strand-level and consensus-level read-depth thresholding, and strand-symmetric CpG merging.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ellis-jace/methylmergeR}
  \item Report bugs at \url{https://github.com/ellis-jace/methylmergeR/issues}
}

}
\author{
\strong{Maintainer}: Jace Ellis \email{ellisjacem@gmail.com}

Authors:
\itemize{
  \item Jace Ellis \email{ellisjacem@gmail.com}
}

}
\keyword{internal}
